feat: optimize costs (remove redis), fix ranking bug, and add seo
