+++ frontend/src/app/robots.ts
@@ -1,0 +1,14 @@
+import { MetadataRoute } from "next";
+
+export default function robots(): MetadataRoute.Robots {
+  const baseUrl = process.env.FRONTEND_URL || "https://todeolho.org";
+
+  return {
+    rules: {
+      userAgent: "*",
+      allow: "/",
+      disallow: ["/api/", "/_next/"],
+    },
+    sitemap: `${baseUrl}/sitemap.xml`,
+  };
+}
